+++ R/jambio-tx.R
@@ -747,9 +747,15 @@ defineDetectedTx <- function
    ## returns transcripts in the same order as each matrix below, which
    ## otherwise has rownames based upon gene and not transcript.
    iRows <- match(rownames(iMatrixTxGrp), tx2geneDF$transcript_id);
-   txExprGrpTx <- shrinkMatrix(rownames(iMatrixTxGrp),
+   if (verbose) {
+      printDebug("defineDetectedTx(): ",
+         "shrinkMatrix Tx names.");
+   }
+   txExprGrpTx <- splicejam::shrinkMatrix(rownames(iMatrixTxGrp),
       groupBy=tx2geneDF[iRows,"gene_name"],
-      shrinkFunc=c, returnClass="matrix");
+      shrinkFunc=c,
+      returnClass="matrix",
+      verbose=TRUE);
    retVals$txExprGrpTx <- txExprGrpTx;
 
 
@@ -871,6 +877,8 @@ defineDetectedTx <- function
 #'    rownames are entries from `groups`.
 #' @param verbose logical indicating whether to print verbose output.
 #'
+#' @import data.table
+#'
 #' @export
 shrinkMatrix <- function
 (x,
@@ -911,16 +919,32 @@ shrinkMatrix <- function
    ## Create DT object
    if (verbose) {
       t1 <- Sys.time();
+      printDebug("shrinkMatrix(): ",
+         "Create DT");
    }
-   DT <- data.table(data.frame(check.names=FALSE, stringsAsFactors=FALSE,
-      x, "groupBy"=groupBy),
+   DF <- data.frame(check.names=FALSE,
+      stringsAsFactors=FALSE,
+      x=x,
+      groupBy=groupBy);
+   DT <- data.table(DF,
       key="groupBy");
    if (verbose) {
       t2 <- Sys.time();
    }
 
    ## Operate on the DT object
-   byDT <- DT[,lapply(.SD, shrinkFunc), by=groupBy];
+   if (verbose) {
+      t1 <- Sys.time();
+      printDebug("shrinkMatrix(): ",
+         "head(DT)");
+      print(head(DT));
+      printDebug("shrinkMatrix(): ",
+         "class(shrinkFunc):",
+         class(shrinkFunc));
+      printDebug("shrinkMatrix(): ",
+         "Create byDT");
+   }
+   byDT <- DT[,lapply(.SD, shrinkFunc), by="groupBy"];
    if (verbose) {
       t3 <- Sys.time();
    }
@@ -932,6 +956,11 @@ shrinkMatrix <- function
       printDebug("shrinkMatrix(): ",
          "Duration for data.table shrinkMatrix: ",
          format(t3-t2));
+   }
+   if (verbose) {
+      t1 <- Sys.time();
+      printDebug("shrinkMatrix(): ",
+         "Create retData data.frame");
    }
    retData <- as(byDT, "data.frame");
    if (returnClass %in% "matrix") {